+++ templates/main.go
@@ -14,12 +14,12 @@ import (
 	"time"
 
 
-
-
 	"<% .Files.Repository %>/internal/database"
-<%if eq (index .Params `fileUploads`) "yes" %>	"<% .Files.Repository %>/internal/file"
+<%if eq (index .Params `fileUploads`) "yes" %>	
+    "<% .Files.Repository %>/internal/file"
 <% end %>
-<%if eq (index .Params `userAuth`) "yes" %>	"<% .Files.Repository %>/internal/auth"
+<%if eq (index .Params `userAuth`) "yes" %>	
+	"<% .Files.Repository %>/internal/auth"
 <% end %>)
 
 const gracefulShutdownTimeout = 10 * time.Second
@@ -54,8 +54,8 @@ func main() {
 
 
 <%if eq (index .Params `fileUploads`) "yes" %>
-	r.HandleFunc("/file/presigned", file.getPresignedUploadURL)
-	r.HandleFunc("/file", file.getPresignedDownloadURL)
+	r.HandleFunc("/file/presigned", file.PresignedUploadURL)
+	r.HandleFunc("/file", file.PresignedDownloadURL)
 <% end %>
