+++ templates/storagecloud/s3service.go
@@ -1,4 +1,4 @@
-package file
+package storagecloud
 
 import (
 	"time"
